+++ tests/integration/rest/test_bitmex.py
@@ -15,7 +15,11 @@ b = Bitmex()
 
 
 def teardown_module(module):
-    asyncio.get_event_loop().run_until_complete(b.shutdown())
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:
+        loop = asyncio.new_event_loop()
+    loop.run_until_complete(b.shutdown())
 
 
 class TestBitmexRest: