+++ tests/integration/rest/test_ftx_us.py
@@ -17,7 +17,11 @@ f = FTXUS(config='config.yaml')
 
 
 def teardown_module(module):
-    asyncio.get_event_loop().run_until_complete(f.shutdown())
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:
+        loop = asyncio.new_event_loop()
+    loop.run_until_complete(f.shutdown())
 
 
 class TestFTXUSRest: